+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* new check `CheckFileExists`
+* new check `CheckFileContainsText`
+
 ## [0.7.1] - 2026-02-22
 
 ### Fixed